[update] move file dir
