+++ r2r/models/sglang_patch/patch.py
@@ -2,9 +2,9 @@ from sglang.srt.managers.scheduler import Scheduler
 from sglang.srt.managers.schedule_batch import Req, ScheduleBatch
 from sglang.srt.layers.attention.flashinfer_backend import FlashInferAttnBackend
 
-from r2r.models.batch_inference.scheduler import __init__ as _init_scheduler, get_next_batch_to_run, get_new_batch_prefill, check_batch_status
-from r2r.models.batch_inference.schedule_batch import __init__ as _init_req, init_next_round_input, reset_for_retract, prepare_for_extend, filter_batch
-from r2r.models.batch_inference.flashinfer_cuda_graph import __init__ as _init_flashinfer_cuda_graph, init_forward_metadata_capture_cuda_graph, init_forward_metadata_replay_cuda_graph, forward_extend
+from r2r.models.sglang_patch.scheduler import __init__ as _init_scheduler, get_next_batch_to_run, get_new_batch_prefill, check_batch_status
+from r2r.models.sglang_patch.schedule_batch import __init__ as _init_req, init_next_round_input, reset_for_retract, prepare_for_extend, filter_batch
+from r2r.models.sglang_patch.flashinfer_cuda_graph import __init__ as _init_flashinfer_cuda_graph, init_forward_metadata_capture_cuda_graph, init_forward_metadata_replay_cuda_graph, forward_extend
 
 FlashInferAttnBackend.__init__ = _init_flashinfer_cuda_graph
 FlashInferAttnBackend.init_forward_metadata_capture_cuda_graph = init_forward_metadata_capture_cuda_graph
