+++ test/test_http_server.py
@@ -5,7 +5,7 @@ import time
 # """
 
 input_ids = [151646, 151646, 151644, 11510, 6556, 362, 7755, 5780, 369, 264, 400, 24, 3, 12, 85526, 20408, 23791, 4227, 323, 17933, 518, 264, 10799, 8061, 26807, 13, 3197, 1340, 22479, 518, 264, 6783, 4628, 315, 400, 82, 3, 40568, 817, 6460, 11, 279, 4227, 4990, 1059, 220, 19, 4115, 11, 2670, 400, 83, 3, 4420, 7391, 304, 279, 10799, 8061, 13, 3197, 1340, 22479, 400, 82, 10, 17, 3, 40568, 817, 6460, 11, 279, 4227, 4990, 1059, 220, 17, 4115, 323, 220, 17, 19, 4420, 11, 2670, 400, 83, 3, 4420, 7391, 304, 279, 10799, 8061, 13, 82610, 362, 7755, 22479, 518, 400, 82, 41715, 37018, 90, 16, 15170, 17, 31716, 40568, 817, 6460, 13, 7379, 279, 1372, 315, 4420, 279, 4227, 4990, 1059, 11, 2670, 279, 400, 83, 3, 4420, 7391, 304, 279, 10799, 8061, 13, 151645, 151648, 198]
-url = "http://0.0.0.0:30005/generate"
+url = "http://0.0.0.0:8000/generate"
 data = {
     # "text": text,
     "input_ids": input_ids,
