+++ test/test_http_server_health.py
@@ -1,10 +1,7 @@
 import requests
 import time
 
-# text = """<｜begin▁of▁sentence｜><｜User｜>Every morning Aya goes for a $9$-kilometer-long walk and stops at a coffee shop afterwards. When she walks at a constant speed of $s$ kilometers per hour, the walk takes her 4 hours, including $t$ minutes spent in the coffee shop. When she walks $s+2$ kilometers per hour, the walk takes her 2 hours and 24 minutes, including $t$ minutes spent in the coffee shop. Suppose Aya walks at $s+\\frac{1}{2}$ kilometers per hour. Find the number of minutes the walk takes her, including the $t$ minutes spent in the coffee shop.<｜Assistant｜><think>
-# """
-
-url = "http://0.0.0.0:30005/health"
+url = "http://0.0.0.0:30000/health"
 
 # Hint: You need to add chat template by yourself if using chat model here
 try:
